backend/requirements.txt
fastapi==0.116.1
Jinja2==3.1.6
pandas==2.3.1
passlib==1.7.4
pydantic==2.11.7
python_jose==3.5.0
Requests==2.32.4
SQLAlchemy==2.0.41
uvicorn==0.34.0
passlib==1.7.4
html5lib==1.1
uuid==1.30
psycopg2==2.9.10